move convolution functions to own file
