+++ nd/filters/tests/test_convolve.py
@@ -1,4 +1,5 @@
-from nd.filters import _convolve, convolve, boxcar, _expand_kernel
+from nd.filters import convolve, boxcar
+from nd.filters.convolve_ import _convolve, _expand_kernel
 from nd.testing import generate_test_dataset
 from nd.io import assemble_complex
 import numpy as np
